moved searchbox into a separate component
